refactor(item model & seed model): seeding real items refactored to seedController
