+++ controllers/itemsController.js
@@ -1,35 +1,6 @@
 const Item = require('./../models/itemModel');
-const scrape = require('./../scrape/get-items');
 
 const itemsController = {};
-
-itemsController.setItemsfromCategory = async (
-  mainCategory,
-  category,
-  limit,
-  offset
-) => {
-  try {
-    const data = await scrape.getCategoryItems(
-      mainCategory,
-      category,
-      limit,
-      offset
-    );
-    const bulk = data.items.map(item => ({
-      item_id: item.id,
-      item_name: item.name,
-      img_url: item.image,
-      amazon_url: item.amazon_link,
-      price: item.price,
-      createdAt: new Date(),
-      updatedAt: new Date()
-    }));
-    await Item.setItemsfromCategory(bulk);
-  } catch (e) {
-    console.log(e); // eslint-disable-line no-console
-  }
-};
 
 itemsController.getRecommended = async (req, res) => {
   try {
