+++ models/itemModel.js
@@ -3,14 +3,6 @@ const _ = require('lodash');
 
 const itemModel = {};
 
-itemModel.setItemsfromCategory = async bulk => {
-  try {
-    await db.Item.bulkCreate(bulk);
-  } catch (error) {
-    console.log(error); // eslint-disable-line no-console
-  }
-};
-
 itemModel.getRecommended = async user_id => {
   const user = await db.User.findOne({
     where: { user_id }
